Separate the GUI again into back- and front-end
